consolidate block editor initializations
