+++ packages/editor/src/components/provider/index.js
@@ -4,7 +4,7 @@
 import { useEffect, useLayoutEffect, useMemo } from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
-import { EntityProvider } from '@wordpress/core-data';
+import { EntityProvider, useEntityBlockEditor } from '@wordpress/core-data';
 import {
 	BlockEditorProvider,
 	BlockContextProvider,
@@ -17,7 +17,6 @@ import { store as noticesStore } from '@wordpress/notices';
  */
 import withRegistryProvider from './with-registry-provider';
 import ConvertToGroupButtons from '../convert-to-group-buttons';
-import usePostContentEditor from './use-post-content-editor';
 import { store as editorStore } from '../../store';
 import useBlockEditorSettings from './use-block-editor-settings';
 
@@ -48,7 +47,11 @@ function EditorProvider( {
 		};
 	}, [] );
 	const { id, type } = __unstableTemplate ?? post;
-	const blockEditorProps = usePostContentEditor( type, id );
+	const [ blocks, onInput, onChange ] = useEntityBlockEditor(
+		'postType',
+		type,
+		{ id }
+	);
 	const editorSettings = useBlockEditorSettings(
 		settings,
 		!! __unstableTemplate
@@ -58,7 +61,6 @@ function EditorProvider( {
 		setupEditor,
 		updateEditorSettings,
 		__experimentalTearDownEditor,
-		__unstableSetupTemplate,
 	} = useDispatch( editorStore );
 	const { createWarningNotice } = useDispatch( noticesStore );
 
@@ -99,13 +101,6 @@ function EditorProvider( {
 		updateEditorSettings( settings );
 	}, [ settings ] );
 
-	// Synchronize the template as it changes
-	useEffect( () => {
-		if ( __unstableTemplate ) {
-			__unstableSetupTemplate( __unstableTemplate );
-		}
-	}, [ __unstableTemplate?.id ] );
-
 	if ( ! isReady ) {
 		return null;
 	}
@@ -115,7 +110,9 @@ function EditorProvider( {
 			<EntityProvider kind="postType" type={ post.type } id={ post.id }>
 				<BlockContextProvider value={ defaultBlockContext }>
 					<BlockEditorProvider
-						{ ...blockEditorProps }
+						value={ blocks }
+						onChange={ onChange }
+						onInput={ onInput }
 						selectionStart={ selectionStart }
 						selectionEnd={ selectionEnd }
 						settings={ editorSettings }
